homework-lesson-10/requirements.txt
deepeval>=2.5.0
pytest>=8.3.0
python-dotenv>=1.1.0
httpx>=0.28.0
